fix darkmode sidebar issue
